+++ client/src/app/products/swap/page.tsx
@@ -13,18 +13,18 @@ import GenericHero from "@/components/generichero.component"
 
 export default function Page() {
   const [{ globalSearchString }] = useContext(GlobalContext)
-  const swapstreamTokenConfig = useFetch("swapstreamtokenconfig", endPoints.swapstreamTokenConfig, HTTPMethods.POST, { searchQuery: globalSearchString })
-  const products = useFetch("get-products", endPoints.getProductConfig, HTTPMethods.POST, { searchQuery: "swapstream" })
-  const selectedProduct = products?.data?.find((product: any) => product.productName === "swapstream")
+  const swapTokenConfig = useFetch("swaptokenconfig", endPoints.swapTokenConfig, HTTPMethods.POST, { searchQuery: globalSearchString })
+  const products = useFetch("get-products", endPoints.getProductConfig, HTTPMethods.POST, { searchQuery: "swap" })
+  const selectedProduct = products?.data?.find((product: any) => product.productName === "swap")
 
   const displayTokens = useCallback(() => {
-    const tokensToDisplay = swapstreamTokenConfig?.data?.map((token: TokenData) => {
+    const tokensToDisplay = swapTokenConfig?.data?.map((token: TokenData) => {
       const genericProductCardProps: GenericProductCardInterface = {
         badgeText: `${token.tokensPerMatic} Tokens/MATIC`,
         className: "decentralized",
         headerText: token.tokenName,
         footerText: token.description,
-        redirectUri: `/products/swapstream/token?tokenAddress=${token.tokenContractAddress}`
+        redirectUri: `/products/swap/token?tokenAddress=${token.tokenContractAddress}`
       }
 
       return <GenericProductCard key={token.tokenContractAddress} genericProductCardProps={genericProductCardProps} />
@@ -32,20 +32,20 @@ export default function Page() {
 
     return (
       <Row className="mt-2 mb-2">
-        <Show when={!!swapstreamTokenConfig?.data?.length}>
+        <Show when={!!swapTokenConfig?.data?.length}>
           <h4 className="text-white">ERC-20 Tokens</h4>
           {tokensToDisplay}
         </Show >
-        <Show when={!swapstreamTokenConfig?.data?.length}>
+        <Show when={!swapTokenConfig?.data?.length}>
           <h4 className="text-white">No ERC-20 Tokens to display</h4>
         </Show>
       </Row>
     )
-  }, [swapstreamTokenConfig?.data])
+  }, [swapTokenConfig?.data])
 
   return (
     <Container>
-      <Show when={!swapstreamTokenConfig.isLoading && !products.isLoading}>
+      <Show when={!swapTokenConfig.isLoading && !products.isLoading}>
         <GenericHero>
           <p className="branding">{selectedProduct?.productName}</p>
           <p className="muted-text mt-3">{selectedProduct?.largeDescription}</p>
@@ -56,7 +56,7 @@ export default function Page() {
         </GenericHero>
         {displayTokens()}
       </Show>
-      <Show when={swapstreamTokenConfig.isLoading || products.isLoading}>
+      <Show when={swapTokenConfig.isLoading || products.isLoading}>
         <Loading />
       </Show>
     </Container>