+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * File naming changes as per new standard
 * Dependency updates & performance improvements
 * Rebranding of Edgepay as Lenstack Pay
+* Rebranding of Swapstream as Lenstack Swap
 
 ### v3.0
 * Created DTO for all controllers wherever required to avoid raw data extraction from request body